Renamed ProfileEditControllers to ProfileControllers
